requirements.txt
firebase_admin==6.6.0
gdown==5.2.0
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.36.0
streamlit==1.41.1
supabase==2.11.0
